fixx 2023-06-01

+++ sql/get_carriera_valida.sql
@@ -1,8 +1,8 @@
 CREATE OR REPLACE FUNCTION get_carriera_valida(id_studenteP varchar)
-RETURNS TABLE (id_insegnamento integer, nome_insegnamento varchar, voto integer) AS $$
+RETURNS TABLE (id_insegnamento integer, nome_insegnamento varchar, voto numeric) AS $$
 BEGIN
     RETURN QUERY
-    SELECT i.id, i."nomeInsegnamento", s.voto
+    SELECT i.id, i."nomeInsegnamento", s.voto::numeric
     FROM insegnamento i 
     INNER JOIN sostiene s ON i.id = s.id_corso
     WHERE s.id_studente = id_studenteP
